backend/requirements.txt
django>=5.0, <6.0
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn
django-environ
pillow
parameterized